refactor #4: User의 List<UserRegion>을 MyRegion이라는 일급 객체로 변경
